Adds validation to ensure the principal exists for all modification operations  (#21)
* :construction: adding validation for conference and session

* :construction: adds validation for speaker routes

+++ src/server/speaker/speaker-router.ts
@@ -6,13 +6,18 @@ import { handleError } from "../../utils/handleError";
 import { SpeakerDetails } from "./speaker";
 import { ConferenceId } from "../conference/conference-id";
 import { SpeakerId } from "./speaker-id";
+import { NoItem } from "../../utils/dynamo/no-item";
 
 const speakerRouter = Router({});
 
 speakerRouter.post("/:conferenceId/speaker", async (req, res, next) => {
   handleError(next, async () => {
-    const conferenceId = req.params.conferenceId;
+    const rawConferenceId = req.params.conferenceId;
+    const conferenceId = ConferenceId.parse(rawConferenceId);
     const context = initialise(config);
+    if (!(await context.conferenceValidator.conferenceExists(conferenceId))) {
+      return res.status(404);
+    }
     const missing = findMissingProperties(req.body, ["name", "bio", "picture"]);
     if (missing.length > 0) {
       return res.status(400).json({
@@ -26,7 +31,7 @@ speakerRouter.post("/:conferenceId/speaker", async (req, res, next) => {
     }
     const { name, bio, picture, socials } = req.body;
     const speakerId = await context.speakerWriteService.addSpeaker(
-      ConferenceId.parse(conferenceId),
+      conferenceId,
       {
         name,
         bio,
@@ -42,13 +47,22 @@ speakerRouter.patch(
   "/:conferenceId/speaker/:speakerId",
   async (req, res, next) => {
     handleError(next, async () => {
-      const { conferenceId, speakerId } = req.params;
+      const { conferenceId: rawConferenceId, speakerId: rawSpeakerId } =
+        req.params;
+      const conferenceId = ConferenceId.parse(rawConferenceId);
+      const speakerId = SpeakerId.parse(rawSpeakerId);
+
       const context = initialise(config);
+      if (
+        !(await context.speakerValidator.speakerExists(conferenceId, speakerId))
+      ) {
+        return res.status(404);
+      }
       const { name, bio, picture, socials } =
         req.body as Partial<SpeakerDetails>;
       const speaker = await context.speakerWriteService.updateSpeaker(
-        ConferenceId.parse(conferenceId),
-        SpeakerId.parse(speakerId),
+        conferenceId,
+        speakerId,
         {
           name,
           bio,
@@ -71,7 +85,13 @@ speakerRouter.get(
         ConferenceId.parse(conferenceId),
         SpeakerId.parse(speakerId),
       );
-      return res.json(speaker);
+      if (speaker.hasValue()) {
+        return res.json(speaker.value);
+      }
+      if (speaker.error instanceof NoItem) {
+        return res.status(404).json(speaker.error.metaData);
+      }
+      throw speaker.error;
     });
   },
 );
@@ -80,11 +100,21 @@ speakerRouter.delete(
   "/:conferenceId/speaker/:speakerId",
   async (req, res, next) => {
     handleError(next, async () => {
-      const { conferenceId, speakerId } = req.params;
+      const { conferenceId: rawConferenceId, speakerId: rawSpeakerId } =
+        req.params;
+      const conferenceId = ConferenceId.parse(rawConferenceId);
+      const speakerId = SpeakerId.parse(rawSpeakerId);
+
       const context = initialise(config);
+      if (
+        !(await context.speakerValidator.speakerExists(conferenceId, speakerId))
+      ) {
+        return res.status(404);
+      }
+
       const speaker = await context.speakerWriteService.deleteSpeaker(
-        ConferenceId.parse(conferenceId),
-        SpeakerId.parse(speakerId),
+        conferenceId,
+        speakerId,
       );
       return res.json(speaker);
     });
@@ -93,11 +123,14 @@ speakerRouter.delete(
 
 speakerRouter.get("/:conferenceId/speaker/", async (req, res, next) => {
   handleError(next, async () => {
-    const { conferenceId } = req.params;
+    const rawConferenceId = req.params.conferenceId;
+    const conferenceId = ConferenceId.parse(rawConferenceId);
     const context = initialise(config);
-    const speakers = await context.speakerReadService.getAllSpeakers(
-      ConferenceId.parse(conferenceId),
-    );
+    if (!(await context.conferenceValidator.conferenceExists(conferenceId))) {
+      return res.status(404);
+    }
+    const speakers =
+      await context.speakerReadService.getAllSpeakers(conferenceId);
     return res.json({ speakers: speakers });
   });
 });
